+++ engine/src/main/java/net/jqwik/engine/properties/DefaultArbitraries.java
@@ -33,7 +33,7 @@ public class DefaultArbitraries {
 		ArrayList<ArbitraryConfigurator> configurators = new ArrayList<>();
 		configurators.add(new SizeConfigurator());
 		configurators.add(new WithNullConfigurator());
-		configurators.add(new UniquenessConfigurator());
+		configurators.add(new UniqueElementsConfigurator());
 
 		// TODO: Remove in 1.5.0
 		configurators.add(new UniqueConfigurator());