+++ engine/src/main/java/net/jqwik/engine/properties/arbitraries/DefaultSetArbitrary.java
@@ -74,8 +74,8 @@ public class DefaultSetArbitrary<T> extends MultivalueArbitraryBase<T, Set<T>> i
 	}
 
 	@Override
-	public SetArbitrary<T> uniqueness(Function<T, Object> by) {
+	public SetArbitrary<T> uniqueElements(Function<T, Object> by) {
 		FeatureExtractor<T> featureExtractor = by::apply;
-		return (SetArbitrary<T>) super.uniqueness(featureExtractor);
+		return (SetArbitrary<T>) super.uniqueElements(featureExtractor);
 	}
 }