+++ engine/src/main/java/net/jqwik/engine/properties/configurators/UniqueElementsConfigurator.java
@@ -2,7 +2,6 @@ package net.jqwik.engine.properties.configurators;
 
 import java.util.*;
 import java.util.function.*;
-import java.util.logging.*;
 import java.util.stream.*;
 
 import net.jqwik.api.*;
@@ -12,14 +11,13 @@ import net.jqwik.api.constraints.*;
 import net.jqwik.api.providers.*;
 import net.jqwik.engine.support.*;
 
-public class UniquenessConfigurator implements ArbitraryConfigurator {
+@SuppressWarnings("unchecked")
+public class UniqueElementsConfigurator implements ArbitraryConfigurator {
 
-	private static final Logger LOG = Logger.getLogger(UniquenessConfigurator.class.getName());
-
-	@SuppressWarnings("unchecked")
+	@SuppressWarnings("OverlyComplexMethod")
 	@Override
 	public <T> Arbitrary<T> configure(Arbitrary<T> arbitrary, TypeUsage targetType) {
-		return targetType.findAnnotation(Uniqueness.class).map(uniqueness -> {
+		return targetType.findAnnotation(UniqueElements.class).map(uniqueness -> {
 			if (arbitrary instanceof ListArbitrary) {
 				return (Arbitrary<T>) configureListArbitrary((ListArbitrary<?>) arbitrary, uniqueness);
 			}
@@ -37,28 +35,28 @@ public class UniquenessConfigurator implements ArbitraryConfigurator {
 			}
 			if (targetType.isAssignableFrom(List.class)) {
 				Arbitrary<List<?>> listArbitrary = (Arbitrary<List<?>>) arbitrary;
-				return (Arbitrary<T>) listArbitrary.filter(l -> isUnique(l, extractor(uniqueness)));
+				return (Arbitrary<T>) listArbitrary.filter(l -> isUnique(l, (Function<Object, Object>) extractor(uniqueness)));
 			}
 			if (targetType.isAssignableFrom(Set.class)) {
 				Arbitrary<Set<?>> setArbitrary = (Arbitrary<Set<?>>) arbitrary;
-				return (Arbitrary<T>) setArbitrary.filter(l -> isUnique(l, extractor(uniqueness)));
+				return (Arbitrary<T>) setArbitrary.filter(l -> isUnique(l, (Function<Object, Object>) extractor(uniqueness)));
 			}
 			if (targetType.isArray()) {
 				Arbitrary<Object[]> arrayArbitrary = (Arbitrary<Object[]>) arbitrary;
-				return (Arbitrary<T>) arrayArbitrary.filter(array -> isUnique(Arrays.asList(array), extractor(uniqueness)));
+				return (Arbitrary<T>) arrayArbitrary.filter(array -> isUnique(Arrays.asList(array), (Function<Object, Object>) extractor(uniqueness)));
 			}
 			if (targetType.isAssignableFrom(Stream.class)) {
 				Arbitrary<Stream<?>> streamArbitrary = (Arbitrary<Stream<?>>) arbitrary;
 				// Since a stream can only be consumed once this is more involved than seems necessary at first glance
 				return (Arbitrary<T>) streamArbitrary.map(s -> s.collect(Collectors.toList()))
-													 .filter(l -> isUnique(l, extractor(uniqueness)))
+													 .filter(l -> isUnique(l, (Function<Object, Object>) extractor(uniqueness)))
 													 .map(Collection::stream);
 			}
 			if (targetType.isAssignableFrom(Iterator.class)) {
 				Arbitrary<Iterator<?>> iteratorArbitrary = (Arbitrary<Iterator<?>>) arbitrary;
 				// Since an iterator can only be iterated once this is more involved than seems necessary at first glance
 				Arbitrary<List<?>> listArbitrary = iteratorArbitrary.map(this::toList);
-				return (Arbitrary<T>) listArbitrary.filter(l -> isUnique(l, extractor(uniqueness))).map(List::iterator);
+				return (Arbitrary<T>) listArbitrary.filter(l -> isUnique(l, (Function<Object, Object>) extractor(uniqueness))).map(List::iterator);
 			}
 			return arbitrary;
 		}).orElse(arbitrary);
@@ -77,34 +75,34 @@ public class UniquenessConfigurator implements ArbitraryConfigurator {
 		return set.size() == list.size();
 	}
 
-	private Arbitrary<?> configureListArbitrary(ListArbitrary<?> arbitrary, Uniqueness uniqueness) {
-		Function<Object, Object> extractor = extractor(uniqueness);
-		return arbitrary.uniqueness(extractor::apply);
+	private <T> Arbitrary<?> configureListArbitrary(ListArbitrary<T> arbitrary, UniqueElements uniqueness) {
+		Function<T, Object> extractor = (Function<T, Object>) extractor(uniqueness);
+		return arbitrary.uniqueElements(extractor);
 	}
 
-	private Arbitrary<?> configureSetArbitrary(SetArbitrary<?> arbitrary, Uniqueness uniqueness) {
-		Function<Object, Object> extractor = extractor(uniqueness);
-		return arbitrary.uniqueness(extractor::apply);
+	private <T> Arbitrary<?> configureSetArbitrary(SetArbitrary<T> arbitrary, UniqueElements uniqueness) {
+		Function<T, Object> extractor = (Function<T, Object>) extractor(uniqueness);
+		return arbitrary.uniqueElements(extractor);
 	}
 
-	private Arbitrary<?> configureArrayArbitrary(ArrayArbitrary<?, ?> arbitrary, Uniqueness uniqueness) {
-		Function<Object, Object> extractor = extractor(uniqueness);
-		return arbitrary.uniqueness(extractor::apply);
+	private <T> Arbitrary<?> configureArrayArbitrary(ArrayArbitrary<T, ?> arbitrary, UniqueElements uniqueness) {
+		Function<T, Object> extractor = (Function<T, Object>) extractor(uniqueness);
+		return arbitrary.uniqueElements(extractor);
 	}
 
-	private Arbitrary<?> configureStreamArbitrary(StreamArbitrary<?> arbitrary, Uniqueness uniqueness) {
-		Function<Object, Object> extractor = extractor(uniqueness);
-		return arbitrary.uniqueness(extractor::apply);
+	private <T> Arbitrary<?> configureStreamArbitrary(StreamArbitrary<T> arbitrary, UniqueElements uniqueness) {
+		Function<T, Object> extractor = (Function<T, Object>) extractor(uniqueness);
+		return arbitrary.uniqueElements(extractor);
 	}
 
-	private Arbitrary<?> configureIteratorArbitrary(IteratorArbitrary<?> arbitrary, Uniqueness uniqueness) {
-		Function<Object, Object> extractor = extractor(uniqueness);
-		return arbitrary.uniqueness(extractor::apply);
+	private <T> Arbitrary<?> configureIteratorArbitrary(IteratorArbitrary<T> arbitrary, UniqueElements uniqueness) {
+		Function<T, Object> extractor = (Function<T, Object>) extractor(uniqueness);
+		return arbitrary.uniqueElements(extractor);
 	}
 
-	private Function<Object, Object> extractor(Uniqueness uniqueness) {
-		Class<? extends Function<Object, Object>> extractorClass = uniqueness.by();
-		return extractorClass.equals(Uniqueness.NOT_SET.class)
+	private Function<?, Object> extractor(UniqueElements uniqueness) {
+		Class<? extends Function<?, Object>> extractorClass = uniqueness.by();
+		return extractorClass.equals(UniqueElements.NOT_SET.class)
 				? Function.identity()
 				// TODO: Create instance in context of test instance.
 				//       This requires an extension of ArbitraryConfiguration interface